refactor(annotations): rename jaxrs package/classes to ws
Rename to framework-neutral names since both javax and jakarta are now supported.
- annotations/jaxrs/ -> annotations/ws/
- JaxRsMappingAnnotation -> WsMappingAnnotation
- JaxRsUrlFormatter -> WsUrlFormatter
- JAXRS_PACKAGE_NAME -> JAVAX_WS_PACKAGE
- JAKARTA_PACKAGE_NAME -> JAKARTA_WS_PACKAGE

고대준 <daejoon.ko@hanatour.com>
